requirements.txt
flask
pytrends
pandas
spacy
openai
python-dotenv